# doc/WatchStoreApi.openapi.3.0.json
{"info":{"title":"WatchStoreApi","description":"","version":"1.0.0","x-project-id":"604ec62e0051000"},"openapi":"3.0.0","components":{"schemas":{"Product":{"type":"object","properties":{"Id":{"type":"integer","default":1001,"examples":[1001],"description":"产品唯一标识符，整数类型\n"},"Name":{"type":"string","default":"纯棉男士T恤","examples":["纯棉男士T恤"],"description":"产品名称"},"Description":{"type":"string","default":"100%纯棉材质，舒适透气，适合日常穿着，有多种颜色可选","examples":["100%纯棉材质，舒适透气，适合日常穿着，有多种颜色可选"],"description":"产品描述"},"ImageUrl":{"type":"string","default":"https://example.com/products/1001/image.jpg","examples":["https://example.com/products/1001/image.jpg"],"description":"产品图片的URL地址，用于展示产品外观\n"},"Material":{"type":"string","default":"100%棉","examples":["100%棉"],"description":"产品使用的材质信息\n"},"Gender":{"type":"string","default":"男","examples":["男"],"description":"产品适用的性别，如男、女、中性等\n"},"Price":{"type":"number","format":"double","default":99.99,"examples":[99.99],"description":"产品价格"},"CategoryId":{"type":"integer","description":"产品的类别Id"}}},"Category":{"type":"object","required":["Name"],"properties":{"Name":{"type":"string"}}}}},"tags":[{"name":"Product","description":"","x-type":"folder","x-url":"","x-target-id":"4ec6c04bd3003"},{"name":"Category","description":"","x-type":"folder","x-url":"","x-target-id":"60e82217d3006"},{"name":"User","description":"","x-type":"folder","x-url":"","x-target-id":"62dc316bd388c"},{"name":"ShoppingCartItem","description":"","x-type":"folder","x-url":"","x-target-id":"85a0a67bd39e2"},{"name":"订单相关","description":"","x-type":"folder","x-url":"","x-target-id":"86c934fbd3b4d"},{"name":"管理员接口","description":"","x-type":"folder","x-url":"","x-target-id":"87d0a717d3f81"}],"paths":{"/api/Products":{"get":{"x-target-id":"4ec9a393d3030","x-protocol":"http","summary":"查询产品","description":"","tags":["Product"],"x-mark-id":"1","x-updated-at":"2026-04-01T16:03:57+08:00","x-updated-user-name":"Feng","parameters":[{"name":"search","in":"query","description":"","required":false,"example":"皮","schema":{"type":"string"}},{"name":"categoryId","in":"query","description":"","required":false,"example":"2","schema":{"type":"number"}},{"name":"material","in":"query","description":"","required":false,"example":"","schema":{"type":"string"}},{"name":"gender","in":"query","description":"","required":false,"example":"","schema":{"type":"string"}},{"name":"minPrice","in":"query","description":"","required":false,"example":"","schema":{"type":"number"}},{"name":"maxPrice","in":"query","description":"","required":false,"example":"","schema":{"type":"number"}},{"name":"pageNumber","in":"query","description":"","required":false,"example":"1","schema":{"type":"number"}},{"name":"pageSize","in":"query","description":"","required":false,"example":"5","schema":{"type":"number"}}],"responses":{"200":{"description":"成功","content":{"application/json":{"schema":{"type":"object","properties":{}},"example":""}}},"404":{"description":"失败","content":{"application/json":{"schema":{"type":"object","properties":{}},"example":""}}}}},"post":{"x-target-id":"4edbf54bd3141","x-protocol":"http","summary":"新增产品","description":"","tags":["Product"],"x-mark-id":"1","x-updated-at":"2026-04-01T14:42:51+08:00","x-updated-user-name":"Feng","requestBody":{"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"Name":{"type":"string","description":"产品的正式名称，用于唯一标识该商品\n","example":"华为WatchGT3"},"Description":{"type":"string","description":"产品核心功能与特点说明，帮助用户快速了解商品\n","example":"华为WatchGT3具备7天长续航和心率监测功能"},"Material":{"type":"string","description":"产品主体及配件使用的材料信息\n","example":"不锈钢表壳+氟橡胶表带"},"Gender":{"type":"string","description":"产品适合的使用性别范围\n","example":"男女通用"},"Price":{"type":"number","description":"产品的市场售价，单位为人民币元\n","example":"1299"},"CategoryId":{"type":"integer","description":"商品所属分类的唯一标识ID\n","example":"2"},"Image":{"type":"string","description":"产品主图的网络地址，用于展示商品外观\n","example":"C:\\Users\\Feng\\Pictures\\【哲风壁纸】抽象艺术-水墨风.png","format":"binary"}},"required":["Description","Material","Gender","Price","CategoryId","Image"]}}}},"responses":{"200":{"description":"成功","content":{"application/json":{"schema":{"type":"object","properties":{}},"example":""}}},"404":{"description":"失败","content":{"application/json":{"schema":{"type":"object","properties":{}},"example":""}}}}}},"/api/Products/{id}":{"put":{"x-target-id":"4efaa9e3d327a","x-protocol":"http","summary":"更新产品","description":"","tags":["Product"],"x-mark-id":"1","x-updated-at":"2026-04-01T14:53:18+08:00","x-updated-user-name":"Feng","parameters":[{"name":"id","in":"path","description":"","required":true,"example":"1","schema":{"type":"integer"}}],"requestBody":{"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"Name":{"type":"string","description":"产品的正式名称，用于唯一标识该商品\n","example":"华为WatchGT3Pro","default":"华为WatchGT3"},"Description":{"type":"string","description":"","example":"华为WatchGT3Pro具备7天长续航和心率监测功能"},"Price":{"type":"string","description":"","example":"9999"},"CategoryId":{"type":"integer","description":"","example":"2"},"Image":{"type":"string","description":"","example":"C:\\Users\\Feng\\Pictures\\【哲风壁纸】清晨-雪山.png","format":"binary"}},"required":["Name","Description","Price","CategoryId","Image"]}}}},"responses":{"200":{"description":"成功","content":{"application/json":{"schema":{"type":"object","properties":{}},"example":""}}},"404":{"description":"失败","content":{"application/json":{"schema":{"type":"object","properties":{}},"example":""}}}}},"delete":{"x-target-id":"4f04162bd3374","x-protocol":"http","summary":"删除产品","description":"","tags":["Product"],"x-mark-id":"1","x-updated-at":"2026-04-01T14:54:00+08:00","x-updated-user-name":"Feng","parameters":[{"name":"id","in":"path","description":"","required":true,"example":"1","schema":{"type":"integer"}}],"responses":{"200":{"description":"成功","content":{"application/json":{"schema":{"type":"object","properties":{}},"example":""}}},"404":{"description":"失败","content":{"application/json":{"schema":{"type":"object","properties":{}},"example":""}}}}}},"/api/Categories":{"post":{"x-target-id":"60e8eaf3d3007","x-protocol":"http","summary":"新增类别","description":"","tags":["Category"],"x-mark-id":"1","x-updated-at":"2026-04-03T08:40:57+08:00","x-updated-user-name":"Feng","parameters":[],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Category"},"example":"{\n\t\"Name\": \"手机\"\n}"}}},"responses":{"200":{"description":"成功","content":{"application/json":{"schema":{"type":"object","properties":{}},"example":""}}},"404":{"description":"失败","content":{"application/json":{"schema":{"type":"object","properties":{}},"example":""}}}}},"get":{"x-target-id":"60fef2b3d30f0","x-protocol":"http","summary":"查询类别","description":"","tags":["Category"],"x-mark-id":"1","x-updated-at":"2026-04-03T08:41:01+08:00","x-updated-user-name":"Feng","responses":{"200":{"description":"成功","content":{"application/json":{"schema":{"type":"object","properties":{}},"example":""}}},"404":{"description":"失败","content":{"application/json":{"schema":{"type":"object","properties":{}},"example":""}}}}}},"/api/Categories/{id}":{"put":{"x-target-id":"60fa335bd3093","x-protocol":"http","summary":"更新类别","description":"","tags":["Category"],"x-mark-id":"1","x-updated-at":"2026-04-01T14:00:28+08:00","x-updated-user-name":"Feng","parameters":[{"name":"id","in":"path","description":"","required":true,"example":"1","schema":{"type":"integer"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Category"},"example":"{\n\t\"Name\": \"陆芳\"\n}"}}},"responses":{"200":{"description":"成功","content":{"application/json":{"schema":{"type":"object","properties":{}},"example":""}}},"404":{"description":"失败","content":{"application/json":{"schema":{"type":"object","properties":{}},"example":""}}}}},"delete":{"x-target-id":"6100473fd3121","x-protocol":"http","summary":"删除类别","description":"","tags":["Category"],"x-mark-id":"1","x-updated-at":"2026-04-01T14:01:38+08:00","x-updated-user-name":"Feng","parameters":[{"name":"id","in":"path","description":"","required":true,"example":"1","schema":{"type":"integer"}}],"responses":{"200":{"description":"成功","content":{"application/json":{"schema":{"type":"object","properties":{}},"example":""}}},"404":{"description":"失败","content":{"application/json":{"schema":{"type":"object","properties":{}},"example":""}}}}}},"/api/Users/register":{"post":{"x-target-id":"62dd312bd388d","x-protocol":"http","summary":"注册用户","description":"","tags":["User"],"x-mark-id":"1","x-updated-at":"2026-04-01T17:01:29+08:00","x-updated-user-name":"Feng","requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"Name":{"type":"string","description":"产品的正式名称，用于唯一标识该商品\n","default":"华为WatchGT3"},"Email":{"type":"string"},"Phone":{"type":"string"},"PasswordHash":{"type":"string"}},"required":["Name","Email","Phone","PasswordHash"]},"example":"{\r\n\t\"Name\": \"王五\",\r\n\t\"Email\": \"wang@gmail.com\",\r\n\t\"Phone\": \"13571858258\",\r\n\t\"PasswordHash\": \"PasswordHash123\"\r\n}"}}},"responses":{"200":{"description":"成功","content":{"application/json":{"schema":{"type":"object","properties":{}},"example":""}}},"404":{"description":"失败","content":{"application/json":{"schema":{"type":"object","properties":{}},"example":""}}}}}},"/api/Users/login":{"post":{"x-target-id":"633c7c33d38f1","x-protocol":"http","summary":"登录","description":"","tags":["User"],"x-mark-id":"1","x-updated-at":"2026-04-03T13:21:32+08:00","x-updated-user-name":"Feng","requestBody":{"content":{"application/json":{"schema":{"type":"object"},"example":"{\r\n    \"Email\": \"wang@gmail.com\",\r\n    \"Password\": \"PasswordHash123\"\r\n}"}}},"responses":{"200":{"description":"成功","content":{"application/json":{"schema":{"type":"object","properties":{}},"example":""}}},"404":{"description":"失败","content":{"application/json":{"schema":{"type":"object","properties":{}},"example":""}}}}}},"/api/ShoppingCartItems/add":{"post":{"x-target-id":"85a1aae7d39e3","x-protocol":"http","summary":"添加商品到购物车","description":"","tags":["ShoppingCartItem"],"x-mark-id":"1","x-updated-at":"2026-04-03T10:25:48+08:00","x-updated-user-name":"Feng","requestBody":{"content":{"application/json":{"schema":{"type":"object"},"example":"{\r\n    \"ProductId\": 8,\r\n    \"Qty\": 2,\r\n    \"UserId\": 3\r\n}"}}},"responses":{"200":{"description":"成功","content":{"application/json":{"schema":{"type":"object","properties":{}},"example":""}}},"404":{"description":"失败","content":{"application/json":{"schema":{"type":"object","properties":{}},"example":""}}}}}},"/api/ShoppingCartItems":{"put":{"x-target-id":"85d733efd3a14","x-protocol":"http","summary":"更新购物车商品数量","description":"","tags":["ShoppingCartItem"],"x-mark-id":"1","x-updated-at":"2026-04-03T09:02:13+08:00","x-updated-user-name":"Feng","parameters":[{"name":"productId","in":"query","description":"","required":true,"example":"11","schema":{"type":"string"}},{"name":"action","in":"query","description":"","required":true,"example":"decrease","schema":{"type":"string"}}],"responses":{"200":{"description":"成功","content":{"application/json":{"schema":{"type":"object","properties":{}},"example":""}}},"404":{"description":"失败","content":{"application/json":{"schema":{"type":"object","properties":{}},"example":""}}}}},"get":{"x-target-id":"860e44d7d3ad1","x-protocol":"http","summary":"获取购物车信息","description":"","tags":["ShoppingCartItem"],"x-mark-id":"1","x-updated-at":"2026-04-03T09:12:01+08:00","x-updated-user-name":"Feng","responses":{"200":{"description":"成功","content":{"application/json":{"schema":{"type":"object","properties":{}},"example":""}}},"404":{"description":"失败","content":{"application/json":{"schema":{"type":"object","properties":{}},"example":""}}}}}},"/api/ShoppingCartItems/remove/{productId}":{"delete":{"x-target-id":"861321b7d3b04","x-protocol":"http","summary":"删除购物车项","description":"","tags":["ShoppingCartItem"],"x-mark-id":"1","x-updated-at":"2026-04-03T09:13:45+08:00","x-updated-user-name":"Feng","parameters":[{"name":"productId","in":"path","description":"","required":true,"example":"11","schema":{"type":"integer"}}],"responses":{"200":{"description":"成功","content":{"application/json":{"schema":{"type":"object","properties":{}},"example":""}}},"404":{"description":"失败","content":{"application/json":{"schema":{"type":"object","properties":{}},"example":""}}}}}},"/api/orders":{"post":{"x-target-id":"86cc60c7d3ba4","x-protocol":"http","summary":"创建订单","description":"","tags":["订单相关"],"x-mark-id":"1","x-updated-at":"2026-04-03T13:21:39+08:00","x-updated-user-name":"Feng","requestBody":{"content":{"application/json":{"schema":{"type":"object"},"example":"{\r\n    \"address\": \"南极\",\r\n    \"userId\": 3\r\n}"}}},"responses":{"200":{"description":"成功","content":{"application/json":{"schema":{"type":"object","properties":{}},"example":""}}},"404":{"description":"失败","content":{"application/json":{"schema":{"type":"object","properties":{}},"example":""}}}}}},"/api/orders/my":{"get":{"x-target-id":"86f39b87d3bd4","x-protocol":"http","summary":"获取订单","description":"","tags":["订单相关"],"x-mark-id":"1","x-updated-at":"2026-04-03T10:14:30+08:00","x-updated-user-name":"Feng","responses":{"200":{"description":"成功","content":{"application/json":{"schema":{"type":"object","properties":{}},"example":""}}},"404":{"description":"失败","content":{"application/json":{"schema":{"type":"object","properties":{}},"example":""}}}}}},"/api/orders/{orderId}/details":{"get":{"x-target-id":"8705cd5bd3c32","x-protocol":"http","summary":"获取订单详情","description":"","tags":["订单相关"],"x-mark-id":"1","x-updated-at":"2026-04-03T10:29:30+08:00","x-updated-user-name":"Feng","parameters":[{"name":"orderId","in":"path","description":"","required":true,"example":"2","schema":{"type":"string"}}],"responses":{"200":{"description":"成功","content":{"application/json":{"schema":{"type":"object","properties":{}},"example":""}}},"404":{"description":"失败","content":{"application/json":{"schema":{"type":"object","properties":{}},"example":""}}}}}},"/api/orders/admin":{"get":{"x-target-id":"8745b507d3cfb","x-protocol":"http","summary":"管理员获取所有订单","description":"","tags":["订单相关"],"x-mark-id":"1","x-updated-at":"2026-04-03T10:49:35+08:00","x-updated-user-name":"Feng","parameters":[{"name":"status","in":"query","description":"","required":true,"example":"","schema":{"type":"string"}},{"name":"startDate","in":"query","description":"","required":true,"example":"","schema":{"type":"string"}},{"name":"endDate","in":"query","description":"","required":true,"example":"","schema":{"type":"string"}},{"name":"user","in":"query","description":"","required":true,"example":"李四","schema":{"type":"string"}},{"name":"pageNumber","in":"query","description":"","required":true,"example":"1","schema":{"type":"integer"}},{"name":"pageSize","in":"query","description":"","required":true,"example":"10","schema":{"type":"integer"}}],"responses":{"200":{"description":"成功","content":{"application/json":{"schema":{"type":"object","properties":{}},"example":""}}},"404":{"description":"失败","content":{"application/json":{"schema":{"type":"object","properties":{}},"example":""}}}}}},"/api/orders/admin/pending":{"get":{"x-target-id":"8761c85bd3e62","x-protocol":"http","summary":"管理员获取所有待处理的订单","description":"","tags":["订单相关"],"x-mark-id":"1","x-updated-at":"2026-04-03T10:49:42+08:00","x-updated-user-name":"Feng","parameters":[{"name":"pagerNumber","in":"query","description":"","required":true,"example":"1","schema":{"type":"integer"}},{"name":"pageSize","in":"query","description":"","required":true,"example":"5","schema":{"type":"integer"}}],"responses":{"200":{"description":"成功","content":{"application/json":{"schema":{"type":"object","properties":{}},"example":""}}},"404":{"description":"失败","content":{"application/json":{"schema":{"type":"object","properties":{}},"example":""}}}}}},"/api/orders/{orderId}/admindetails":{"get":{"x-target-id":"87771d23d3ee8","x-protocol":"http","summary":"管理员获取订单详情","description":"","tags":["订单相关"],"x-mark-id":"1","x-updated-at":"2026-04-03T10:50:47+08:00","x-updated-user-name":"Feng","parameters":[{"name":"orderId","in":"path","description":"","required":true,"example":"2","schema":{"type":"integer"}}],"responses":{"200":{"description":"成功","content":{"application/json":{"schema":{"type":"object","properties":{}},"example":""}}},"404":{"description":"失败","content":{"application/json":{"schema":{"type":"object","properties":{}},"example":""}}}}}},"/api/orders/{orderId}/status":{"put":{"x-target-id":"879591ebd3f21","x-protocol":"http","summary":"管理员修改订单状态","description":"","tags":["订单相关"],"x-mark-id":"1","x-updated-at":"2026-04-03T10:59:45+08:00","x-updated-user-name":"Feng","parameters":[{"name":"orderStatus","in":"query","description":"","required":true,"example":"Completed","schema":{"type":"string"}},{"name":"orderId","in":"path","description":"","required":true,"example":"1","schema":{"type":"integer"}}],"responses":{"200":{"description":"成功","content":{"application/json":{"schema":{"type":"object","properties":{}},"example":""}}},"404":{"description":"失败","content":{"application/json":{"schema":{"type":"object","properties":{}},"example":""}}}}}},"/api/admin/revenue":{"get":{"x-target-id":"87d29a3fd3fad","x-protocol":"http","summary":"获取收入","description":"","tags":["管理员接口"],"x-mark-id":"1","x-updated-at":"2026-04-03T11:15:38+08:00","x-updated-user-name":"Feng","parameters":[{"name":"range","in":"query","description":"","required":true,"example":"yearly","schema":{"type":"string"}}],"responses":{"200":{"description":"成功","content":{"application/json":{"schema":{"type":"object","properties":{}},"example":""}}},"404":{"description":"失败","content":{"application/json":{"schema":{"type":"object","properties":{}},"example":""}}}}}},"/api/admin/dashboard":{"get":{"x-target-id":"89722f47d300e","x-protocol":"http","summary":"管理员Dashboard","description":"","tags":["管理员接口"],"x-mark-id":"1","x-updated-at":"2026-04-03T13:08:38+08:00","x-updated-user-name":"Feng","responses":{"200":{"description":"成功","content":{"application/json":{"schema":{"type":"object","properties":{}},"example":""}}},"404":{"description":"失败","content":{"application/json":{"schema":{"type":"object","properties":{}},"example":""}}}}}}},"servers":[{"url":"https://localhost:7220","description":"默认环境","x-env-id":"","variables":{}}]}